add tests for SDK getEventsTup
