+++ src/contract/templates/simplecontract.h
@@ -24,10 +24,10 @@ class SimpleContract : public DynamicContract {
 
   public:
     /// Event for when the name changes.
-    void nameChanged(const EventParam<std::string, false>& name) { this->emitEvent(__func__,  std::make_tuple(name)); }
+    void nameChanged(const EventParam<std::string, true>& name) { this->emitEvent(__func__,  std::make_tuple(name)); }
 
     /// Event for when the value changes.
-    void valueChanged(const EventParam<uint256_t, true>& value) { this->emitEvent(__func__, std::make_tuple(value)); }
+    void valueChanged(const EventParam<uint256_t, false>& value) { this->emitEvent(__func__, std::make_tuple(value)); }
 
     /// Event for when the name and value change. used for testing json abi generation
     void nameAndValueChanged(const EventParam<std::string, true>& name, const EventParam<uint256_t, true>& value) {
